Move pure pursue into commands folder
