+++ src/main/java/robot/subsystems/drivetrain/commands/PurePursue.java
@@ -1,8 +1,11 @@
-package robot.subsystems.drivetrain.pure_pursuit;
+package robot.subsystems.drivetrain.commands;
 
+import com.sun.xml.internal.bind.v2.runtime.reflect.opt.Const;
 import edu.wpi.first.wpilibj.command.Command;
 import robot.Robot;
 import robot.subsystems.drivetrain.Drivetrain;
+import robot.subsystems.drivetrain.pure_pursuit.*;
+import robot.subsystems.drivetrain.pure_pursuit.Constants;
 
 /**
  *
