+++ src/main/java/robot/subsystems/drivetrain/pure_pursuit/Constants.java
@@ -3,7 +3,7 @@ package robot.subsystems.drivetrain.pure_pursuit;
 public class Constants {
     //DRIVING CONSTANTS
     public final static double LOOKAHEAD_DISTANCE = 0.3; //in meters
-    final static double TRACK_WIDTH = 3;
+    public final static double TRACK_WIDTH = 3;
 
     //PATH GENERATION CONSTANTS
     static final double SPACING_BETWEEN_WAYPOINTS = 0.1524; //meters
@@ -19,5 +19,5 @@ public class Constants {
     public final static double kA = 0;
     public final static double kP = 0;
 
-    final static double STOP_SPEED_THRESH = 0.1; //the speed the robot could stop at the end of the path.
+    public final static double STOP_SPEED_THRESH = 0.1; //the speed the robot could stop at the end of the path.
 }
